requirements.txt
python-dotenv
langchain-openai
langchain-core
langchain-classic
langchain-text-splitters
langchain-huggingface
langchain-community
requests
tavily
uvicorn
psycopg2-binary
sqlalchemy
pgvector
numpy
fastapi
python-multipart
redis
sentence-transformers
torch
huggingface-hub
--index-url https://mirrors.aliyun.com/pypi/simple
